% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layered_bdm.R
\name{layered_bdm}
\alias{layered_bdm}
\title{Title}
\usage{
layered_bdm(x)
}
\arguments{
\item{A}{matrix or SpatRaster}
}
\value{
A numeric vector
}
\description{
Title
}
\examples{
data(mini_array)
layered_bdm(mini_array)
}
