% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_array}
\alias{mini_array}
\title{Small matrix for testing}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 8 columns.
}
\usage{
mini_array
}
\description{
A small matrix for testing the layered_bdm function
}
\keyword{datasets}
